Initial commit.
This is the game Haunted House, the old adventure game from the 1983
Usborne book "Write Your Own Adventure Programs For Your Microcomputer"
by Jenny Tyler and Les Howarth.

This is a direct translation from the BASIC code which was meant to run
on an early-80s microcomputer, for which 16K of RAM was considered an
insane quantity of memory.  The writers wrote it to be able to run on a
VIC 20.  It shows.