Little things.
Added a bit of documentation explaining what an object's flag does.

Fixed the grammar in the "You broke the wall" message.

Actually double the score for getting back to the start with everything.

Since objects have this weird 1-based array, fix the bit of code
that tells you what you see when you enter a room.